+++ ITZZA/ITZZA/Screen/Community/Cell/TVC/PostTVC.swift
@@ -92,38 +92,7 @@ extension PostTVC {
     }
     
     func bindButtonAction() {
-        didTapLikeButton()
-        didTapBookmarkButton()
-    }
-    
-    func didTapLikeButton() {
-        footerView.likeButton.rx.tap
-             .scan(false) { lastState, newState in
-                 !lastState
-             }
-             .subscribe(onNext: {
-                 self.footerView.likeButton.setImageToggle($0, UIImage(systemName: "heart")!, UIImage(systemName: "heart.fill")!)
-                 self.footerView.likeCnt.text = self.setButtonCnt($0, self.footerView.likeCnt.text!)
-             })
-             .disposed(by: bag)
-    }
-    
-    func didTapBookmarkButton() {
-        footerView.bookmarkButton.rx.tap
-             .scan(false) { lastState, newState in
-                 !lastState
-             }
-             .subscribe(onNext: {
-                 self.footerView.bookmarkButton.setImageToggle($0, UIImage(systemName: "bookmark")!, UIImage(systemName: "bookmark.fill")!)
-             })
-             .disposed(by: bag)
-    }
-    
-    func setButtonCnt(_ state: Bool, _ lastCnt: String) -> String {
-        if state {
-            return String(Int(lastCnt)! + 1)
-        } else {
-            return String(Int(lastCnt)! - 1)
-        }
+        footerView.didTapLikeButton()
+        footerView.didTapBookmarkButton()
     }
 }
